requirements.txt
pydantic-settings
pandas
numpy
streamlit
plotly
faiss-cpu
sentence-transformers
google-generativeai
